Mans sync torus orient algorithm
