+++ orientation/sync_torus/sync_mans.go
@@ -4,10 +4,13 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
-	"math"
-	"math/rand"
 
 	"github.com/krzysztof-turowski/distributed-framework/lib"
+)
+
+/* TESTS */
+const (
+	testCases int = 100
 )
 
 /* NODE LABELS */
@@ -18,11 +21,11 @@ type DirectionType int
 
 /* BOUNDS */
 const (
-	msg_bound  int           = 16
+	msgBound   int           = 16
 	directions DirectionType = 4
 )
 
-/* NODE TYPES */
+/* NODE TYPES ONLY USED IN SKETCHES */
 type NodeStateType int
 
 const (
@@ -49,11 +52,6 @@ type Message struct {
 	ToGo     int
 }
 
-/* TESTS */
-const (
-	testCases int = 100
-)
-
 /* STATE FUNCTIONS */
 type StateNodeTorus struct {
 	// counter of received messages
@@ -69,12 +67,15 @@ type StateNodeTorus struct {
 	BoundaryClosed bool
 }
 
-func newState(dim int) *StateNodeTorus {
-	var st StateNodeTorus
-	st.Count = 0
-	st.H1 = make(map[DirectionType]NodeLabelType)
-	st.H2 = make(map[DirectionType][]NodeLabelType)
-	st.Pending = make([][]*Message, dim)
+func newState() *StateNodeTorus {
+	st := StateNodeTorus{
+		Count:          0,
+		H1:             make(map[DirectionType]NodeLabelType),
+		H2:             make(map[DirectionType][]NodeLabelType),
+		Pending:        make([][]*Message, directions),
+		State:          passive,
+		BoundaryClosed: false,
+	}
 	for i := range st.Pending {
 		st.Pending[i] = make([]*Message, 0)
 	}
@@ -82,8 +83,6 @@ func newState(dim int) *StateNodeTorus {
 	for i := DirectionType(0); i < directions; i++ {
 		st.H2[i] = make([]NodeLabelType, 0)
 	}
-	st.State = passive
-	st.BoundaryClosed = false
 	return &st
 }
 
@@ -156,23 +155,9 @@ func addPendingMessage(index int, st *StateNodeTorus, msg *Message) {
 	st.Pending[index] = append(st.Pending[index], msg)
 }
 
-func clearPendingMessages(st *StateNodeTorus) {
-	for i := range st.Pending {
-		st.Pending[i] = make([]*Message, 0)
-	}
-}
-
-func clearAllPendingMessages(vertices []lib.Node) {
-	for _, v := range vertices {
-		st := getState(v)
-		clearPendingMessages(st)
-		setState(v, st)
-	}
-}
-
 /* TORUS ORIENTATION METHODS */
 func initializeNode(v lib.Node) bool {
-	st := newState(v.GetOutChannelsCount())
+	st := newState()
 	log.Println("Node", v.GetIndex(), "started orientation protocol")
 
 	msg := newMessageOrient(one, NodeLabelType(v.GetIndex()))
@@ -228,7 +213,7 @@ func orient(v lib.Node) bool {
 	st := getState(v)
 	finish := false
 
-	if st.Count == msg_bound {
+	if st.Count == msgBound {
 		finish = true
 	}
 
@@ -252,45 +237,6 @@ func run(v lib.Node) {
 	log.Println("Node", v.GetIndex(), "finished.")
 }
 
-/* BUILD */
-func square_check(x int) bool {
-	var int_root int = int(math.Sqrt(float64(x)))
-	return (int_root * int_root) == x
-}
-
-func BuildSynchronizedTorus(n int) ([]lib.Node, lib.Synchronizer) {
-	if n <= 0 {
-		panic("Size cannot be <= 0.")
-	}
-	if !square_check(n) {
-		panic("n is not a perfect square.")
-	}
-
-	if int(math.Sqrt(float64(n))) <= 4 {
-		panic("Sqrt(n) should be greater than 4.")
-	}
-
-	log.Println("Building torus size", n)
-	adjacencyList := make([][]int, n)
-
-	for i := range adjacencyList {
-		adjacencyList[i] = make([]int, 0)
-	}
-
-	side := int(math.Sqrt(float64(n)))
-
-	for i := 0; i < side; i++ {
-		for j := 0; j < side; j++ {
-			adjacencyList[i*side+j] = append(adjacencyList[i*side+j], i*side+((j-1+side)%side))
-			adjacencyList[i*side+j] = append(adjacencyList[i*side+j], i*side+((j+1+side)%side))
-			adjacencyList[i*side+j] = append(adjacencyList[i*side+j], (((i-1+side)%side)*side + j))
-			adjacencyList[i*side+j] = append(adjacencyList[i*side+j], (((i+1+side)%side)*side + j))
-		}
-	}
-
-	return lib.BuildSynchronizedGraphFromAdjacencyList(adjacencyList, lib.GetRandomGenerator())
-}
-
 /* CHECK */
 func getNodeInd(vertices []lib.Node, id NodeLabelType) int {
 	for i := range vertices {
@@ -302,7 +248,7 @@ func getNodeInd(vertices []lib.Node, id NodeLabelType) int {
 	return -1
 }
 
-func getLinkFrom(v lib.Node, id NodeLabelType) DirectionType {
+func getLinkFromId(v lib.Node, id NodeLabelType) DirectionType {
 	st := getState(v)
 	for i, k := range st.H1 {
 		if NodeLabelType(k) == id {
@@ -313,77 +259,6 @@ func getLinkFrom(v lib.Node, id NodeLabelType) DirectionType {
 	return -1
 }
 
-func checkWithoutRunning(vertices []lib.Node, initiatorNode int) {
-	dMin := 1
-	d := rand.Intn(len(vertices)) + dMin
-	distance := 0
-	currentNode := vertices[initiatorNode]
-	prevNode := vertices[initiatorNode]
-	st := getState(currentNode)
-	k, r := findPerpendicularAny(st)
-	handrail := st.H1[k]
-	toGo := d
-	soFar := 0
-	distance++
-	nextNodeId := st.H1[r]
-
-	log.Println("Node:", currentNode.GetIndex(), "initiator.")
-	log.Println("d:", d)
-
-	for distance < 4*d {
-		prevNode = currentNode
-		currentNode = vertices[getNodeInd(vertices, nextNodeId)]
-		soFar++
-		r, k = findPerpendicularVal(getState(currentNode), handrail, getLinkFrom(currentNode, NodeLabelType(prevNode.GetIndex())))
-		if soFar != toGo {
-			st = getState(currentNode)
-			handrail = st.H1[k]
-			s := findParallelTo(st, r)
-			nextNodeId = st.H1[s]
-		} else {
-			st = getState(currentNode)
-			handrail = st.H1[r]
-			soFar = 0
-			nextNodeId = st.H1[k]
-		}
-		log.Println("Next node", nextNodeId, ".")
-		distance++
-	}
-
-	if nextNodeId != NodeLabelType(vertices[initiatorNode].GetIndex()) {
-		panic("Boundary not closed!!!")
-	}
-}
-
-func multipleTests(vertices []lib.Node) {
-	numOfVer := len(vertices)
-	initiatiorNode := rand.Intn(numOfVer)
-	for i := 0; i < testCases; i++ {
-		checkWithoutRunning(vertices, initiatiorNode)
-	}
-
-}
-
-/* RUN */
-func start(vertices []lib.Node, synchronizer lib.Synchronizer) (int, int) {
-	for _, v := range vertices {
-		go run(v)
-	}
-	synchronizer.Synchronize(2)
-	synchronizer.GetStats()
-	return synchronizer.GetStats()
-}
-
-func Run(n int) {
-	vertices, sync := BuildSynchronizedTorus(n)
-	msgs, rounds := start(vertices, sync)
-	fmt.Println("ORIENTATION")
-	fmt.Println("Rounds:", rounds, "Messages:", msgs)
-	fmt.Println("TESTING CORECTNESS")
-	multipleTests(vertices)
-}
-
-/* MY SKETCHES FOR RUNTIME BOUNDARY, MAY BE WRONG BUT USEFUL */
 func emptyIntersect(h21 []NodeLabelType, h22 []NodeLabelType) bool {
 	if len(h22) == 0 || len(h21) == 0 {
 		return true
@@ -428,7 +303,7 @@ func findParallelTo(st *StateNodeTorus, link DirectionType) DirectionType {
 	return s
 }
 
-func findPerpendicularVal(st *StateNodeTorus, handrail NodeLabelType, link DirectionType) (DirectionType, DirectionType) {
+func findPerpendicularTo(st *StateNodeTorus, handrail NodeLabelType, link DirectionType) (DirectionType, DirectionType) {
 	// links
 	r, k := link, DirectionType(-1)
 	for i := DirectionType(0); i < directions; i++ {
@@ -447,98 +322,73 @@ func findPerpendicularVal(st *StateNodeTorus, handrail NodeLabelType, link Direc
 	return r, k
 }
 
-func markTerritory(v lib.Node, st *StateNodeTorus) {
-	for i := DirectionType(0); i < DirectionType(v.GetInChannelsCount()); i++ {
-		msgin := receiveMessage(v, int(i))
-		if msgin != nil {
-			if msgin.Type == token {
-				if st.State == initiator {
-					log.Println("Node:", v.GetIndex(), "boundary closed.")
-					st.BoundaryClosed = true
-				} else {
-					sofar := msgin.SoFar + 1
-					r, k := findPerpendicularVal(st, msgin.Handrail, i)
-					if sofar != msgin.ToGo {
-						log.Println("Node", v.GetIndex(), "straight.")
-						handrail := st.H1[k]
-						msgout := newMessageHandrail(NodeLabelType(v.GetIndex()), handrail, sofar, msgin.ToGo)
-						s := findParallelTo(st, r)
-						addPendingMessage(int(s), st, msgout)
-					} else {
-						log.Println("Node", v.GetIndex(), "turn.")
-						handrail := st.H1[r]
-						msgout := newMessageHandrail(NodeLabelType(v.GetIndex()), handrail, 0, msgin.ToGo)
-						addPendingMessage(int(k), st, msgout)
-					}
-					st.BoundaryClosed = true
-				}
-			} else {
-				// not reachable
-			}
-		}
-	}
-}
-
-func mark(v lib.Node) bool {
-	st := getState(v)
-	finish := st.BoundaryClosed
-
-	markTerritory(v, st)
-	sendPendingMessages(v, st)
-	setState(v, st)
-
-	return finish
-}
-
-func startMarking(v lib.Node, d int) bool {
-	log.Println("Node:", v.GetIndex(), "starting to mark territory.")
-	st := getState(v)
+func checkWithoutRunning(vertices []lib.Node, initiatorNode int) {
+	gen := lib.GetRandomGenerator()
+	dMin := 1
+	d := gen.Int()%len(vertices) + dMin
+	distance := 0
+	currentNode := vertices[initiatorNode]
+	prevNode := vertices[initiatorNode]
+	st := getState(currentNode)
 	k, r := findPerpendicularAny(st)
 	handrail := st.H1[k]
-	msgout := newMessageHandrail(NodeLabelType(v.GetIndex()), handrail, 0, d)
-	addPendingMessage(int(r), st, msgout)
-	sendPendingMessages(v, st)
-	setState(v, st)
-	return false
-}
+	toGo := d
+	soFar := 0
+	nextNodeId := st.H1[r]
 
-func check(v lib.Node, d int) {
-	finish := false
-	st := getState(v)
-	v.StartProcessing()
-	if st.State == initiator {
-		finish = startMarking(v, d)
-	} else {
-		finish = mark(v)
-	}
-	v.FinishProcessing(finish)
-	distance := 1
+	log.Println("Node:", currentNode.GetIndex(), "initiator.")
+	log.Println("d:", d)
+
+	distance++
+
 	for distance < 4*d {
-		v.StartProcessing()
-		finish = mark(v)
-		v.FinishProcessing(finish)
+		prevNode = currentNode
+		currentNode = vertices[getNodeInd(vertices, nextNodeId)]
+		soFar++
+		r, k = findPerpendicularTo(getState(currentNode), handrail, getLinkFromId(currentNode, NodeLabelType(prevNode.GetIndex())))
+		if soFar != toGo {
+			st = getState(currentNode)
+			handrail = st.H1[k]
+			s := findParallelTo(st, r)
+			nextNodeId = st.H1[s]
+		} else {
+			st = getState(currentNode)
+			handrail = st.H1[r]
+			soFar = 0
+			nextNodeId = st.H1[k]
+		}
+		log.Println("Next node", nextNodeId, ".")
 		distance++
 	}
-}
 
-func checkAll(vertices []lib.Node, synchronizer lib.Synchronizer) (int, int) {
-	clearAllPendingMessages(vertices)
-
-	initiatorNode := 0
-	dBound := int(math.Sqrt(float64(len(vertices))))
-	dMin := 2
-	d := rand.Intn(dBound-dMin) + dMin
-	st := getState(vertices[initiatorNode])
-	st.State = initiator
-	setState(vertices[initiatorNode], st)
-
-	for _, v := range vertices {
-		go check(v, d)
-	}
-	synchronizer.Synchronize(0)
-	synchronizer.GetStats()
-	if !getState(vertices[initiatorNode]).BoundaryClosed {
+	if nextNodeId != NodeLabelType(vertices[initiatorNode].GetIndex()) {
 		panic("Boundary not closed!!!")
 	}
+}
+
+func multipleTests(vertices []lib.Node) {
+	gen := lib.GetRandomGenerator()
+	initiatiorNode := gen.Int() % len(vertices)
+	for i := 0; i < testCases; i++ {
+		checkWithoutRunning(vertices, initiatiorNode)
+	}
+
+}
+
+/* RUN */
+func start(vertices []lib.Node, synchronizer lib.Synchronizer) (int, int) {
+	for _, v := range vertices {
+		go run(v)
+	}
+	synchronizer.Synchronize(0)
 	return synchronizer.GetStats()
+}
+
+func Run(n int) {
+	vertices, sync := lib.BuildSynchronizedTorus(n)
+	msgs, rounds := start(vertices, sync)
+	fmt.Println("ORIENTATION")
+	fmt.Println("Rounds:", rounds, "Messages:", msgs)
+	fmt.Println("TESTING CORECTNESS")
+	multipleTests(vertices)
 }
