+++ text/1193274/tex/references.bib
@@ -1,0 +1,15 @@
+@article{mans,
+    author  = "Bernard Mans",
+    title   = "Optimal Distributed Algorithms in Unlabeled Tori and Chordal Rings",
+    year    = "1997",
+    journal = "JOURNAL OF PARALLEL AND DISTRIBUTED COMPUTING"
+}
+
+@article{peterson,
+    author  = "Gary Lynn Peterson",
+    title   = "Efficient Algorithms for Elections
+in Meshes and Complete Networks",
+    year    = "1985",
+    journal = "TR 140",
+
+}
